refactor(core): move recorder generation into workflows
