+++ packages/core/src/ai-model/index.ts
@@ -12,30 +12,24 @@ export {
 } from './connectivity/index';
 export { systemPromptToLocateElement } from './prompt/llm-locator';
 export {
-  generatePlaywrightTest,
-  generatePlaywrightTestStream,
-} from './prompt/playwright-generator';
-export {
   convertRecordLogIntoMarkdown,
   createRecorderMarkdownReplayPrompt,
+  generatePlaywrightTest,
+  generatePlaywrightTestStream,
   generateRecorderMarkdownReplay,
-} from './prompt/markdown-generator';
-export type { RecorderMarkdownGenerationInput } from './prompt/markdown-generator';
-export { generateRecorderSessionMetadata } from './prompt/recorder-metadata-generator';
-export type {
-  RecorderGeneratedMetadata,
-  RecorderMetadataGenerationInput,
-} from './prompt/recorder-metadata-generator';
-export {
+  generateRecorderSessionMetadata,
   generateRecorderYamlTest,
   generateRecorderYamlTestStream,
   generateYamlTest,
   generateYamlTestStream,
-} from './prompt/yaml-generator';
+} from './workflows/recorder-generation';
 export type {
+  RecorderMarkdownGenerationInput,
+  RecorderGeneratedMetadata,
+  RecorderMetadataGenerationInput,
   RecorderYamlGenerationInput,
   YamlGenerationOptions,
-} from './prompt/yaml-generator';
+} from './workflows/recorder-generation';
 
 export type { ChatCompletionMessageParam } from 'openai/resources/index';
