+++ packages/core/src/ai-model/workflows/recorder-generation/common.ts
