+++ packages/core/src/ai-model/workflows/recorder-generation/index.ts
@@ -1,0 +1,26 @@
+export {
+  generatePlaywrightTest,
+  generatePlaywrightTestStream,
+} from './playwright';
+export {
+  convertRecordLogIntoMarkdown,
+  createRecorderMarkdownReplayPrompt,
+  generateRecorderMarkdownReplay,
+} from './markdown';
+export { generateRecorderSessionMetadata } from './metadata';
+export {
+  generateRecorderYamlTest,
+  generateRecorderYamlTestStream,
+  generateYamlTest,
+  generateYamlTestStream,
+} from './yaml';
+
+export type { RecorderMarkdownGenerationInput } from './markdown';
+export type {
+  RecorderGeneratedMetadata,
+  RecorderMetadataGenerationInput,
+} from './metadata';
+export type {
+  RecorderYamlGenerationInput,
+  YamlGenerationOptions,
+} from './yaml';
